connected front end to back end of blog and routine. Add post, add comment, remove post, add routine, remove routine. Separated front end to components
