--- client/src/styles/CreatePost.css
@@ -1,21 +1,0 @@
-/* h1 {
-    text-align: center;
-    font-family: var(--font);
-    color: white;
-    font-size: 25px;
-    margin-top: 20px;
-    margin-bottom: 20px;
-}
-
-form {
-    color: white;
-}
-
-p {
-    text-align: center;
-    font-family: var(--font);
-    color: white;
-    font-size: 25px;
-    margin-top: 20px;
-    margin-bottom: 20px;
-} */
