+++ client/src/utils/queries.js
@@ -7,8 +7,26 @@ export const QUERY_POSTS = gql`
             title
             text
             author
-            likes
+            likes {
+              username
+            }
             createdAt
+            comments {
+              _id
+              commentText
+              commentAuthor
+              commentCreatedAt
+          }
+        }
+    }
+`
+
+export const QUERY_POST = gql`
+    query getPost($postId: ID!) {
+        post(postId: $postId){
+            _id
+            title
+            text
         }
     }
 `
@@ -20,18 +38,19 @@ export const QUERY_USER_POST = gql`
             title
             text
             author
-            likes
+            likes {
+              username
+            }
             createdAt
             comments {
                 _id
                 commentText
                 commentAuthor
-                createdAt
+                commentCreatedAt
             }
         }
     }
 `
-
 
 export const QUERY_PROFILES = gql`
   query allProfiles {
@@ -91,11 +110,14 @@ export const QUERY_USER = gql`
             text
             author
             createdAt
+            likes {
+              username
+            }
             comments {
                 _id
                 commentText
                 commentAuthor
-                createdAt
+                commentCreatedAt
             }
       }
         profile {
@@ -127,11 +149,14 @@ export const QUERY_ME = gql`
         text
         author
         createdAt
+        likes {
+              username
+            }
         comments {
           _id
           commentText
           commentAuthor
-          createdAt
+          commentCreatedAt
         }
       }
         profile {
